pyproject.toml
[project]
name = "prob_inf_marl"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "jamesrj_desktop", email = "james.rudd.jones@gmail.com"},
    {name = "james_dell", email = "james.rudd.jones@gmail.com"},
]
dependencies = [
    "absl-py>=2.1.0",
    "rlax==0.1.2",
#    "stable-baselines3==2.0.0a1",
#    "gymnasium[accept-rom-license,atari,other]==0.28.1",
#    "flax>=0.7.4",
    "optax>=0.2.2",
    "wandb>=0.17.1",
#    "ale-py>=0.8.1",
#    "dm-haiku>=0.0.12",
    "bsuite>=0.3.5",
#    "scipy==1.8.0",
    "flashbax>=0.1.2",
    "lxm3==0.4.2",
#    "tensorflow[and-cuda]>=2.17.0",
    "gymnax>=0.0.8",
    "jaxmarl>=0.0.2",
    "pygame>=2.6.1",
#    "jax[cuda12_pip]>=0.4.18",
#    "jax[cpu]==0.4.23",
#    "jax==0.5.0",
    "jax==0.6.0",
    "flax==0.10.6",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

 [build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

 [[tool.pdm.source]]
 name = "jax_cuda"
 url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
 verify_ssl = true
 type = "find_links"

                                               [tool.pdm.scripts]
test.cmd = "pytest -n auto --maxprocesses 8 --dist loadfile"
# test.env = { JAX_PLATFORM_NAME = "cpu", JAX_DISABLE_MOST_OPTIMIZATIONS = "1" }
lint.shell = "ruff . && black --check ."
fmt.shell = "ruff --fix . && black ."

                                               [tool.pdm.dev-dependencies]
dev = [
    "lxm3>=0.4.0",
    "ruff>=0.1.11",
    "black>=23.12.1",
    "pytest>=7.4.4",
    "pytest-xdist>=3.5.0",
    "pre-commit>=3.6.0",
]

# Ruff configuration
# https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 120
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
]
ignore = ["E731"]

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
force-single-line = true
force-sort-within-sections = true
single-line-exclusions = ["typing"]

# Black configuration
# https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.github
  | \.venv
  | docs
  | third_party
  | scratch
)/
'''

